Fixing bugs. SWI-DSP works fine again. Now it has simplified functors, supports n channels and has a better data module

+++ swidata/blobidOperations.cpp
@@ -89,7 +89,7 @@ PREDICATE(get_blob_size, 2)
 		return false;
 	}
 
-	return A2 = PlTerm(double(ch1->size()));		
+	return A2 = PlTerm(long(ch1->size()));		
 }
 
 /**
